+++ common/src/main/java/de/the_build_craft/remote_player_waypoints_for_xaero/common/waypoints/PlayerPosition.java
@@ -29,7 +29,7 @@ import net.minecraft.client.player.AbstractClientPlayer;
  *
  * @author ewpratten
  * @author Leander Knüttel
- * @version 29.08.2025
+ * @version 31.08.2025
  */
 public class PlayerPosition extends Position {
     public GameProfile gameProfile;
@@ -40,13 +40,13 @@ public class PlayerPosition extends Position {
         this.world = world;
     }
 
-    public PlayerPosition(String name, float x, float y, float z, String world) {
+    public PlayerPosition(String name, double x, double y, double z, String world) {
         super(name, x, y, z, name, "players");
         this.world = world;
     }
 
     public PlayerPosition(AbstractClientPlayer player) {
-        super(player.getGameProfile().getName(), (float) player.getX(), (float) player.getY(), (float) player.getZ(), player.getGameProfile().getName(), "players");
+        super(player.getGameProfile().getName(), player.getX(), player.getY(), player.getZ(), player.getGameProfile().getName(), "players");
         this.world = "";
         this.gameProfile = player.getGameProfile();
     }
